requirements.txt
pytest
sqlalchemy-cratedb